feat(Task#78896): migrate database

+++ src/main/java/cinemas/repositories/UserRepository.java
@@ -2,5 +2,5 @@ package cinemas.repositories;
 
 import cinemas.models.User;
 
-public interface UserRepository extends BaseRepository<User, Long> {
+public interface UserRepository extends BaseRepository<User, Integer> {
 }
